Update keywords
I decided to use one table (otherwise we would need many SQLAlchemy models, it's not convenient). On keywords in text issue: I created column "keywords" in the table. We can store keywords in this column and then search

So, here's how it'll work:
1. Create database of all California bills (from site)
2. For keyword we need, make search query with these keywords on site. In database search bills from site search results and append keyword to their "keyword" column.
No need for multiple tables with identical columns and keeping full bill text. And search results will be same as on site

update_keywords function is already in file